폴더 구조 변경 - Components 폴더 내 영역 별 폴더 추가 - 폴더 구조 및 User 컴포넌트 구조 변경
